Remove strip_goto
This was meant to resolve the problem that, when building Xen, the
mkreloc tool complains about the goto-cc section not having the same
relative virtual address in build artifacts. The problem, however, needs
to be resolved in goto-cc directly as the object file cannot be edited
using objcopy at that point anymore.