smatch: use cgcc instead of smatch
To make most out of the smatch tooling, use the smatch compiler
wrapper cgcc instead of calling smatch directly. This way, tooling
inside smatch is used as well, which might reduce failures in the
future. At the moment, this is mostly handling command line arguments,
similarly to what we do in our wrapper.