plain: do not extend as and ld calls
In case extra arguments for the compiler are given, do not forward
them to the called native tool if it is an assembler, as the parameters
might not be known to both the compiler and the assembler.

Likewise, ld does not share all parameters of gcc, so that the extra
arguments will also not be passed to ld.