infer: add backend
This backend allows to wrap compiler calls with "infer capture", to
collect the files that then can be fed into "infer analyze". As we might
have to deal with different compilers, we split the typical single call
into multiple calls, and add the "--continue" flag, to be able to trace
all compiler calls independently and in parallel. Finally, we point the
infer analyze call to the used output directory, so that a report can be
created. From this report, we generate a variant in GCC style, which can
then be processed further.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>