analyze: allow to run independent of build
For now, analysis could only be run in case the build was also run.
However, there are cases where the two steps should be split. To achieve
this, we add another flag to the script, that allows to skip the build
command, and furthermore asks for using the results of a previous call
already.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>