plain: add compilation database
When compiling a project with one-line-scan, an interesting output
format is compilation databases, which can be consumed by other tools to
run further analysis on the target software. To make these databases
more accessible, dump this format when using the plain wrapper as well.

To keep track of dependencies in compilation, we add a nanosecond based
timestamp to each JSON blob that is created per compilation command.
When aggregating all JSON blobs, this time stamp is used to determine
the order.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>