infer: filter gcc params
When projects are compiled with gcc, they can use CLI parameters for the
compiler that infer, as well as the internally used clang, do not
understand. To make sure we can still analyze this kind of projects,
these parameters have to be dropped from the compiler calls, or
rewritten into a clang equivalent.

This commit adds a first set of gcc parameters that have to be dropped
in order to run infer capture successfully.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>