infer: auto-fix capture errors
In case infer capture fails due to compiler arguments being used that
clang cannot handle, an error will be reported. We can catch this error
and drop the responsible argument from all compile commands that are
used for capturing. This allows to increase capturing coverage at the
cost of higher run time.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>