one-line-cr-bot: initial commit
This script compiles a project with cppcheck and infer for 2 specified
commits, the current HEAD commit, as well as the upstream commit.

In case the findings for the two commits are different, the script will
present them accordingly.

To tell which introduced defects are actually new, we use the new python
script to match new defects independently of their line number
information. This allows to either report only new signatures, or also
report findings whose line information changed.

The script can be configured to just present the defects, or fail on new
defects, as well as ignore defects with line changes.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>