comparison: identify new gcc msgs
We sometimes want to be able to identify messages that have been
introduced newly, and follow the gcc message style "file:line: ...".

This script allows to filter for messages that are new in a given file
in comparison to another file, by comparing new files independently of
their line number, and next combines them with the line number again to
actually present them.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>