wrapper: allow extra compiler
Certain build systems might rely on other compilers than gcc, g++, clang
or clang++, and the combinations we allow by specifying prefix and suffix.
This change introduces the support of an environment variable, which allows
to specify further target compiler. When setting a value for the variable
OLS_TARGET_COMPILER, its content will be considered as compilers to be
wrapped. If you want to track multiple additional compiler, separate their
names with a space.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>