cppcheck: use env for customization
Similarly to other tools, allow CppCheck to be configured via
environment variable. As the analysis already happens during the
compilation step, the environment variable needs to be evaluated in the
compiler wrapper. Other tools implement this as part of the evaluation
script.

As we are now able to configure the analysis, the aggressive default
setup is reduced, as users can specify these command line options
themselves. The parameters that are dropped from the cppcheck invocation
are:

--enable=style --enable=performance --enable=information
--enable=portability --inconclusive

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>