github: add comment script
This script allows to add a comment for a pull request, namely the pull
request that is currently running a workflow.

Authentication is done via the GITHUB_TOKEN environment variable, that
should be forwarded from github secrets.

The comments to be posted should come from a findings file. The user can
choose to only comment a summary.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>