docker: add a minimum dockerfile
This adds a Dockerfile which can be used as a base image in other projects to
support one-line-scan in a docker container.

A sample invocation to build the "one-line-scan" image, could be:

```
docker build -t one-line-scan tools/
```

Later on, this can be used in other projects as a base image.

Signed-off-by: Martin Mazein <amazein@amazon.de>