cbmc: do not require goto-diff
The tool goto-diff is not installed in PATH when the cbmc package is
instally via the system. Hence, do not rely on this tool being present.
However, if it is present, use it as before this change.

By this, we basically lose the recursive include analysis for the CBMC
backend.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>