cr-bot: use pull_request
The bot that consumes code changes and analyzes them should not have
access to a GITHUB TOKEN that provides write access. Otherwise, info
about this token could be leaked via some channel.

Hence, do recomment to NOT use this feature for now.

A workaround is a split model where the currentl action writes the
reported defects as artifacts, and has access to the read-only GITHUB
TOKEN. Next, another scheduled workflow of the upstream repository will
check for new PRs and their artifacts, and will post them to the PRs as
comment. The second workflow is static, and owned by the upstream
repository, so a PR cannot leak the secret GITHUB TOKEN.

Signed-off-by: Norbert Manthey <nmanthey@amazon.de>